pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Compares two configuration files and shows a difference."
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["PyYAML"]

[project.scripts]
gendiff = "scripts.gendiff:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = ["ruff", "pytest"]

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[dependency-groups]
dev = [
    "isort>=6.0.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.3",
]
